Fix(Field): do not destroy dropdown if used by another container

+++ inc/field.class.php
@@ -343,6 +343,11 @@ class PluginFieldsField extends CommonDBChild
     // phpcs:ignore PSR1.Methods.CamelCapsMethodName
     public function pre_deleteItem()
     {
+        /**
+         * @var \DBmysql $DB
+         */
+        global $DB;
+
         //retrieve search option ID to clean DiplayPreferences
         $container_obj = new PluginFieldsContainer();
         $container_obj->getFromDB($this->fields['plugin_fields_containers_id']);
@@ -388,7 +393,21 @@ class PluginFieldsField extends CommonDBChild
             'items_id' => $this->fields['id'],
         ]);
 
-        if ($this->fields['type'] === 'dropdown') {
+        //load all container and check if another use this fields
+        $container_obj = new PluginFieldsContainer();
+        $all_container = $container_obj->find();
+
+        $use_by_another = false;
+        foreach ($all_container as $container_fields) {
+            foreach (json_decode($container_fields['itemtypes']) as $itemtype) {
+                $classname = PluginFieldsContainer::getClassname($itemtype, $container_fields['name']);
+                if ($DB->fieldExists(getTableForItemType($classname), $this->fields['name'])) {
+                    $use_by_another = true;
+                }
+            }
+        }
+
+        if ($this->fields['type'] === 'dropdown' && !$use_by_another) {
             return PluginFieldsDropdown::destroy($this->fields['name']);
         }
